Move StatusHandler into RequestHandler family
